feat: Add initial app configuration and loading state
Introduces a new `AppConfig` interface and a basic loading screen to `index.html`.
This change also refactors the dependency import for React and React DOM in `index.html` to be more consistent.
The `INITIAL_CONFIG` constant provides default values for social media links and contact information, which will be used to configure the application's external links.